+++ CHANGES.md
@@ -271,7 +271,13 @@
 
 ## V1.7 - Beta
 
-* ...
+* Removed void clearMessageHandlers()
+* Added std::vector<std::function<LRESULT(MSG msg)>> getMessageHandlers() - allows erasure of a custom handler.
+* Added void disableSomeClicks(ws::Hue hue) - Choose what color should represent emptiness and any click will pass through that color as if nothing is there.
+* Added void disableAnyClicks() - force all clicks to pass through this window as if it does not exist.
+
+
+
 
 
 # New major features to be added.